Retrieve AWS last used date for volume from CloudTrail events
This is a bit of a hack and it only works on one volume at a time, yet
it is a valid solution nonetheless. Note that it seems LookupEvents is
only taking the first lookup attribute for filtering, so we need to
iterate over the results later.

Also, AWS only report events for the past 90 days, anything greater
than that it will be discarded.

Signed-off-by: Leandro López (inkel) <leandro.lopez@grafana.com>