Add fake provider
This is a simple provider that returns a random list of unused
disks. The goal of this package is to provide mock implementations of
`unused.Provider` and `unused.Disk` that are useful for testing
purposes.

Because the RNG is always created with the same seeds we can expect
deterministic output, making things easier for E2E testing.

Note that this package can only be compiled when pasing `-tags=fake`
to the `go` command, otherwise it will fail. This will be more evident
in upcoming commits.